LPS-164401 Move scroll to AppLayout
