+++ modules/apps/site-navigation/site-navigation-admin-web/src/main/resources/META-INF/resources/site_navigation_menu_editor/components/Menu.js
@@ -12,57 +12,16 @@
  * details.
  */
 
-import {useSessionState} from '@liferay/layout-content-page-editor-web';
-import {debounce} from 'frontend-js-web';
-import React, {useEffect, useRef} from 'react';
+import React from 'react';
 
-import {useConstants} from '../contexts/ConstantsContext';
 import {useItems} from '../contexts/ItemsContext';
 import {MenuItem} from './MenuItem';
 
 export function Menu() {
 	const items = useItems();
 
-	const {portletNamespace} = useConstants();
-
-	const [scrollPosition, setScrollPosition] = useSessionState(
-		`${portletNamespace}_scrollPosition`,
-		0
-	);
-
-	const containerRef = useRef(null);
-	const scrollPositionRef = useRef(scrollPosition);
-
-	useEffect(() => {
-		const containerElement = containerRef.current;
-
-		if (!containerElement) {
-			return;
-		}
-
-		if (scrollPositionRef.current) {
-			containerElement.scrollY = scrollPositionRef.current;
-		}
-
-		const handleScroll = debounce(() => {
-			setScrollPosition(containerElement.scrollY);
-		}, 300);
-
-		containerElement.addEventListener('scroll', handleScroll, {
-			passive: true,
-		});
-
-		return () => {
-			containerElement.removeEventListener('scroll', handleScroll);
-		};
-	}, [setScrollPosition]);
-
 	return (
-		<div
-			className="container ml-lg-auto ml-sm-0 p-3 pt-4"
-			ref={containerRef}
-			role="list"
-		>
+		<div className="container ml-lg-auto ml-sm-0 p-3 pt-4" role="list">
 			{items.map((item) => (
 				<MenuItem item={item} key={item.siteNavigationMenuItemId} />
 			))}
